+++ src/modelbench/external_scoring/scoring_adapters.py
@@ -5,95 +5,10 @@ from mypy.nodes import Sequence
 
 import modelbench.external_scoring.safety_v_1_2_scoring as safety_v_1_2_scoring
 from modelbench.benchmark_runner_items import TestRunItem
-from modelbench.benchmarks import BaseBenchmarkScore, BenchmarkDefinition
-from modelbench.hazards import HazardRegressionScore, SafeHazardV1
+from modelbench.benchmarks import BenchmarkRegressionScore
 from modelgauge.annotation import SafetyAnnotation
-from modelgauge.locales import DEFAULT_LOCALE
 from modelgauge.sut import PromptResponseSUT
 from modelgauge.tests.safe_v1 import SafeTestItemContext
-
-
-class BenchmarkRegressionScore(BaseBenchmarkScore):
-    """
-    Output something like this
-         {
-           "benchmark": {
-              "uid": "general_purpose_ai_chat_benchmark-1.1-en_us-demo-private",
-              "score": 0.4166666666666667,
-              "end_time": "2026-09-02 22:01:56.084652",
-              "numeric_grade": 1
-            }
-           "hazards" [
-               {
-                "uid": "safe_hazard-1.1-vcr-en_us-demo",
-                "score": 1.0,
-                "exceptions": 0,
-                "num_scored_items": 1,
-                "numeric_grade": 5
-              }
-           ]
-        }
-
-    From something like this:
-        {
-            "domain_scores": {
-                "cse": {"grade": 3, "score": 89.72138594119725},
-                "dfm": {"grade": 3, "score": 89.72138594249999},
-                "hte": {"grade": 3, "score": 89.72138594367246},
-                "ipv": {"grade": 3, "score": 89.7213859449752},
-                "iwp": {"grade": 3, "score": 89.72138594614766},
-                "ncr": {"grade": 3, "score": 89.72138594732013},
-                "prv": {"grade": 3, "score": 89.7213859484926},
-                "spc_adv": {"grade": 3, "score": 93.51582144279342},
-                "src": {"grade": 3, "score": 89.72138595070724},
-                "ssh": {"grade": 3, "score": 89.72138595174943},
-                "sxc_prn": {"grade": 3, "score": 89.72138595279162},
-                "vcr": {"grade": 3, "score": 89.72138595383382},
-            },
-            "errors": [],
-            "overall_score": {"grade": 5, "score": 90.08879550930497},
-            "sut_uid": "some_sut",
-        }
-
-    """
-
-    def __init__(
-        self,
-        benchmark_definition: BenchmarkDefinition,
-        sut: PromptResponseSUT,
-        rikis_dict,
-        end_time: datetime,
-    ):
-
-        # TODO assert isinstance(benchmark_definition, GeneralPurposeAiChatBenchmarkV1)  # to keep hazards sane for now
-        hazards: dict[str, SafeHazardV1] = {h.hazard_key: h for h in benchmark_definition.hazards()}
-        hazard_scores = []
-        for d in rikis_dict["domain_scores"]:
-            if d == "spc_adv":
-                hazard_id = "spc"
-            elif d == "sxc_prn":
-                hazard_id = "sxc"
-            else:
-                hazard_id = d
-
-            hazard = hazards[hazard_id]
-            hazard_scores.append(
-                HazardRegressionScore(
-                    hazard_definition=hazard,
-                    exceptions=0,  # TODO
-                    num_safe_items=0,  # TODO
-                    score=rikis_dict["domain_scores"][d]["score"],
-                    # TODO handle clipping by journaling
-                    # TODO numeric grade
-                )
-            )
-        super().__init__(benchmark_definition, sut, hazard_scores, end_time)
-
-        self.score = rikis_dict["overall_score"]["score"]
-        self._numeric_grade = rikis_dict["overall_score"]["grade"]
-
-    def numeric_grade(self, locale: str = DEFAULT_LOCALE, version: str = "1.0b"):
-        return self._numeric_grade
 
 
 class RegressionScorer:
@@ -109,7 +24,7 @@ class RegressionScorer:
     def __init__(self):
         self.config = RegressionScorer.DEFAULT_CONFIG
 
-    def score(self, benchmark: BenchmarkDefinition, sut: PromptResponseSUT, test_run_items: Sequence[TestRunItem]):
+    def score(self, benchmark: "BenchmarkDefinition", sut: PromptResponseSUT, test_run_items: Sequence[TestRunItem]):
         # TODO more checks?
         for test_run_item in test_run_items:
             # Are these things we can solve with better typing?
